requirements.txt
aiohttp==3.11.11
beautifulsoup4==4.12.3
chromadb==0.4.24
langchain==0.1.12
langchain-community==0.0.29
ollama==0.1.5
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1
tiktoken==0.8.0
lxml==5.3.0
nest_asyncio
pyarrow 